Refactor - Renaming and re-organizing
