+++ VirtualFinland.UserAPI/src/VirtualFinland.UsersAPI/Activities/User/Operations/CreateSearchProfile.cs
@@ -4,6 +4,7 @@ using Swashbuckle.AspNetCore.Annotations;
 using VirtualFinland.UserAPI.Data;
 using VirtualFinland.UserAPI.Exceptions;
 using VirtualFinland.UserAPI.Helpers;
+using VirtualFinland.UserAPI.Helpers.Swagger;
 
 namespace VirtualFinland.UserAPI.Activities.User.Operations;
