+++ VirtualFinland.UserAPI/src/VirtualFinland.UsersAPI/Helpers/Configurations/IIdentityProviderConfig.cs
@@ -1,4 +1,4 @@
-namespace VirtualFinland.UserAPI.Helpers;
+namespace VirtualFinland.UserAPI.Helpers.Configurations;
 
 public interface IIdentityProviderConfig
 {
