+++ VirtualFinland.UserAPI/src/VirtualFinland.UsersAPI/Helpers/Configurations/TestBedIdentityProviderConfig.cs
@@ -1,6 +1,6 @@
 using System.Text.Json.Nodes;
 
-namespace VirtualFinland.UserAPI.Helpers;
+namespace VirtualFinland.UserAPI.Helpers.Configurations;
 
 public class TestBedIdentityProviderConfig : IIdentityProviderConfig
 {
