+++ VirtualFinland.UserAPI/src/VirtualFinland.UsersAPI/Helpers/Swagger/SwaggerIgnoreAttribute.cs
@@ -1,0 +1,6 @@
+namespace VirtualFinland.UserAPI.Helpers.Swagger;
+
+public class SwaggerIgnoreAttribute : Attribute
+{
+
+}
