--- VirtualFinland.UserAPI/src/VirtualFinland.UsersAPI/Helpers/SwaggerIgnoreAttribute.cs
@@ -1,8 +1,0 @@
-namespace VirtualFinland.UserAPI.Helpers;
-
-using System;
-
-public class SwaggerIgnoreAttribute : Attribute
-{
-
-}
